feat: rebase and integrate Explore section into DevDocs page
