+++ src/routes/docs/+page.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
   import DocsMain from './components/DocsMain.svelte';
+  import Explore from './components/Explore.svelte';
 </script>
 
 <DocsMain />
+<Explore classes="mt-4" />
